% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{berger_tls}
\alias{berger_tls}
\title{Calculate true anomaly and true longitude}
\usage{
berger_tls(n, N)
}
\arguments{
\item{n}{Numeric, day of year.}

\item{N}{Numeric, daus in a year.}
}
\value{
True anomaly and true longitude for a given day.
}
\description{
Calculate true anomaly and true longitude
}
\references{
Berger, A. L. (1978), Long term variations of daily insolation
and quaternary climatic changes, J. Atmos. Sci., 35, 2362-2367.
}
