% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evap.R
\name{calc_daily_evap}
\alias{calc_daily_evap}
\title{Calculate daily evaporation fluxes}
\usage{
calc_daily_evap(
  lat,
  n,
  elv = 0,
  y = 0,
  sf = 1,
  tc = 23,
  sw = 1,
  ke = 0.0167,
  keps = 23.44,
  komega = 283,
  kw = 0.26
)
}
\arguments{
\item{lat}{double, decimal degrees.}

\item{n}{double, day of year.}

\item{elv}{double, elevation, m A.S.L. Default: \code{0}.}

\item{y}{double, year. Default: \code{0}.}

\item{sf}{double, fraction of sunshine hours. Default: \code{1}.}

\item{tc}{double, mean daily air temperature, degrees C.
Default: \code{23.0}.}

\item{sw}{double, evaporative supply rate, mm/hr. Default: \code{1.0}.}

\item{ke}{double, eccentricity of earth's orbit, 2000CE (Berger, 1978).
Default: \code{0.01670}.}

\item{keps}{double, obliquity of earth's elliptic, 2000CE (Berger, 1978).
Default: \code{23.44}.}

\item{komega}{double, lon. of perihelion, degrees, 2000CE (Berger, 1978).
Default: \code{283}}

\item{kw}{double, PET entrainment, \eqn{(1 + kw) * EET}
(Priestley-Taylor, 1972). Default: \code{0.26}.}
}
\value{
Returns a \code{list} object with the following variables:
\itemize{
\item nu_deg ............ true anomaly, degrees
\item lambda_deg ........ true longitude, degrees
\item dr ................ distance factor, unitless
\item delta_deg ......... declination angle, degrees
\item hs_deg ............ sunset angle, degrees
\item ra_j.m2 ........... daily extraterrestrial radiation, J/m^2
\item tau ............... atmospheric transmittivity, unitless
\item ppfd_mol.m2 ....... daily photosyn photon flux density, mol/m^2
\item hn_deg ............ net radiation hour angle, degrees
\item rn_j.m2 ........... daily net radiation, J/m^2
\item rnn_j.m2 .......... daily nighttime net radiation, J/m^2
\item econ_m3.j ......... water to energy conversion, m^3/J
\item cond_mm ........... daily condensation, mm
\item eet_mm ............ daily equilibrium evapotranspiration, mm
\item pet_mm ............ daily potential evapotranspiration, mm
\item hi_deg ............ intersection hour angle, degrees
\item aet_mm ............ daily actual evapotranspiration, mm
}
}
\description{
This function calculates daily radiation, condensation, and evaporation
fluxes.
}
\references{
Berger, A.L., 1978. Long-term variations of daily insolation and Quaternary
climatic changes. Journal of Atmospheric Sciences, 35(12), pp.2362-2367.
doi:10.1175/1520-0469(1978)035<2362:LTVODI>2.0.CO;2

Priestley, C.H.B. and Taylor, R.J., 1972. On the assessment of surface heat
flux and evaporation using large-scale parameters. Monthly weather review,
100(2), pp.81-92. doi:10.1175/1520-0493(1972)100<0081:OTAOSH>2.3.CO;2
}
