% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{calc_daily_solar}
\alias{calc_daily_solar}
\title{Calculate daily solar radiation fluxes}
\usage{
calc_daily_solar(
  lat,
  n,
  elv = 0,
  y = 0,
  sf = 1,
  tc = 23,
  ke = 0.0167,
  keps = 23.44,
  komega = 283,
  kalb_sw = 0.17,
  kalb_vis = 0.03,
  kb = 0.2,
  kc = 0.25,
  kd = 0.5,
  kfFEC = 2.04,
  kGsc = 1360.8
)
}
\arguments{
\item{lat}{double, decimal degrees.}

\item{n}{double, day of year.}

\item{elv}{double, elevation, m A.S.L. Default: \code{0}.}

\item{y}{double, year. Default: \code{0}.}

\item{sf}{double, fraction of sunshine hours. Default: \code{1}.}

\item{tc}{double, mean daily air temperature, degrees C.
Default: \code{23.0}.}

\item{ke}{double, eccentricity of earth's orbit, 2000CE (Berger, 1978).
Default: \code{0.01670}.}

\item{keps}{double, obliquity of earth's elliptic, 2000CE (Berger, 1978).
Default: \code{23.44}.}

\item{komega}{double, lon. of perihelion, degrees, 2000CE (Berger, 1978).
Default: \code{283}}

\item{kalb_sw}{double, shortwave albedo, Eq. (10) (Federer, 1968).}

\item{kalb_vis}{double, visible light albedo, Eq. (17) (Sellers, 1985).}

\item{kb}{double, empirical constant, Eq. (13) (Linacre, 1968).}

\item{kc}{double, cloudy transmittivity, Eq. (12) (Linacre, 1968).}

\item{kd}{double, angular coefficient of transmittivity, Eq. (12)
(Linacre, 1968).}

\item{kfFEC}{double, flux-to-energy conversion, Eq. (17) (Meek et al., 1984).}

\item{kGsc}{double, solar constant, Eq. (2) (Kopp and Lean, 2011).}
}
\value{
Returns a \code{list} object with the following variables:
\itemize{
\item nu_deg ............ true anomaly, degrees
\item lambda_deg ........ true longitude, degrees
\item dr ................ distance factor, unitless
\item delta_deg ......... declination angle, degrees
\item hs_deg ............ sunset angle, degrees
\item ra_j.m2 ........... daily extraterrestrial radiation, J/m^2
\item tau ............... atmospheric transmittivity, unitless
\item ppfd_mol.m2 ....... daily photosyn. photon flux density, mol/m^2
\item hn_deg ............ net radiation hour angle, degrees
\item rn_j.m2 ........... daily net radiation, J/m^2
\item rnn_j.m2 .......... daily nighttime net radiation, J/m^2
}
}
\description{
This function calculates daily solar radiation fluxes.
}
\references{
Berger, A.L., 1978. Long-term variations of daily insolation and Quaternary
climatic changes. Journal of Atmospheric Sciences, 35(12), pp.2362-2367.
doi:10.1175/1520-0469(1978)035<2362:LTVODI>2.0.CO;2

Federer, C.A., 1968. Spatial variation of net radiation, albedo and surface
temperature of forests. Journal of Applied Meteorology and Climatology, 7(5),
pp.789-795. doi:10.1175/1520-0450(1968)007<0789:SVONRA>2.0.CO;2

Kopp, G. and Lean, J.L., 2011. A new, lower value of total solar irradiance:
Evidence and climate significance. Geophys. Res. Lett. 38, L01706.
doi:10.1029/2010GL045777

Linacre, E.T., 1968. Estimating the net-radiation flux. Agricultural
meteorology, 5(1), pp.49-63. doi:10.1016/0002-1571(68)90022-8

Meek, D.W., Hatfield, J.L., Howell, T.A., Idso, S.B. and Reginato, R.J.,
1984. A generalized relationship between photosynthetically active radiation
and solar radiation 1. Agronomy journal, 76(6), pp.939-945.
doi:10.2134/agronj1984.00021962007600060018x

Sellers, P.J., 1985. Canopy reflectance, photosynthesis and transpiration,
International Journal of Remote Sensing, 6:8, 1335-1372,
doi:10.1080/01431168508948283
}
