% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{dcos}
\alias{dcos}
\title{Calculate cosine of an angle}
\usage{
dcos(d, pir = pi/180)
}
\arguments{
\item{d}{Numeric, angle in degrees.}
}
\value{
Cosine of an angle.
}
\description{
Calculates the cosine of an angle (d) given in degrees.
}
\references{
C. Johnson, Theoretical Physicist, Univ of Chicago
\itemize{
\item 'Equation of Time' URL: \url{https://mb-soft.com/public3/equatime.html}
\item Javascript URL: \url{https://mb-soft.com/believe/txx/astro22.js}
}
}
\keyword{internal}
