% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evap.R
\name{density_h2o}
\alias{density_h2o}
\title{Calculate density of water at 1 atm, g/cm^3}
\usage{
density_h2o(tc, pa)
}
\arguments{
\item{tc}{double, air temperature, degrees C.}

\item{pa}{double, atm pressure, Pa.}
}
\value{
double, kg/m^3.
}
\description{
This function calculates the temperature and pressure dependent density of
pure water.
}
\references{
Chen, C.T., Fine, R.A. and Millero, F.J., 1977. The equation of state of
pure water determined from sound speeds. The Journal of Chemical Physics,
66(5), pp.2142-2144. doi:10.1063/1.434179
}
\keyword{internal}
