% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_genmodel_jar.R
\name{download_genmodel_jar}
\alias{download_genmodel_jar}
\title{Download h2o-genmodel.jar file}
\usage{
download_genmodel_jar()
}
\description{
This function downloads a copy of \code{h2o-genmodel.jar} to a local package app
data folder, which is then automatically copied to each mojo file path when
\code{splash()} is called. This function is automatically run the first time the
package is used. This function can be called directly to internally update
the \code{h2o-genmodel.jar} file. The \code{h2o-genmodel.jar} file is accessed from the
local \code{h2o} package installation by launching a temporary instance of an
\code{h2o} server on port 426. If a more recent version of the \code{h2o-genmodel.jar}
file is required, please update your locally installed \code{h2o} package.
}
