% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{dsin}
\alias{dsin}
\title{Calculate sine of an angle}
\usage{
dsin(d, pir = pi/180)
}
\arguments{
\item{d}{Numeric, angle in degrees.}
}
\value{
Sine of an angle.
}
\description{
Calculates the sine of an angle (d) given in degrees.
}
\keyword{internal}
