% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evap.R
\name{enthalpy_vap}
\alias{enthalpy_vap}
\title{Calculate enthalpy of vaporization}
\usage{
enthalpy_vap(tc)
}
\arguments{
\item{tc}{double, air temperature, degrees C.}
}
\value{
double, J/kg.
}
\description{
This function calculates the temperature-dependent enthalpy of vaporization
(latent heat of vaporization).
}
\references{
Eq. 8, Henderson‐Sellers, B., 1984. A new formula for latent heat of
vaporization of water as a function of temperature. Quarterly Journal of the
Royal Meteorological Society, 110(466), pp.1186-1190.
\doi{10.1002/qj.49711046626}
}
\keyword{internal}
