% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{julian_day}
\alias{julian_day}
\title{Calculate Julian day}
\usage{
julian_day(y, m, i)
}
\arguments{
\item{y}{double, year.}

\item{m}{double, month.}

\item{i}{double, day of month.}
}
\value{
double, Julian day.
}
\description{
This function converts a date in the Gregorian calendar
to a Julian day number (i.e., a method of consecutive
numbering of days---does not have anything to do with
the Julian calendar!)
}
\details{
\itemize{
\item valid for dates after -4712 January 1 (i.e., jde >= 0)
}
}
\references{
Meeus, J. 1991. Chapter 7 "Julian Day". Astronomical Algorithms.
Willmann-Bell.
}
