% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evap.R
\name{psychro}
\alias{psychro}
\title{Calculate psychrometric constant}
\usage{
psychro(tc, pa)
}
\arguments{
\item{tc}{double, air temperature, degrees C.}

\item{pa}{double, atm pressure, Pa.}
}
\value{
double, Pa/K.
}
\description{
This function calculates the temperature and pressure dependent
psychrometric constant.
}
\references{
Allen, R.G., Pereira, L.S., Raes, D. and Smith, M., 1998. Crop
evapotranspiration-Guidelines for computing crop water requirements-FAO
Irrigation and drainage paper 56. Food and Agriculture Organization of the
United Nations, Rome, 300(9), p.D05109. Available:
http://www.fao.org/docrep/x0490e/x0490e07.htm
}
\keyword{internal}
