% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_txt}
\alias{read_txt}
\title{Read plain text file}
\usage{
read_txt(my_data, fname, var, y = -1)
}
\arguments{
\item{my_data}{List same as the output from \code{\link{read_csv}}.}

\item{fname}{String, file name.}

\item{var}{String, variable name.}

\item{y}{Numeric, year.}
}
\value{
List with the following properties:
\describe{
\item{$file_name}{File name.}
\item{$sf}{Sunshine fraction.}
\item{$tair}{Air temperature.}
\item{$pn}{Precipitation.}
\item{$num_lines}{Number of data points.}
\item{$year}{Year of data.}
}
}
\description{
Reads plain text file (no header) of one of the input arrays.
}
