% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash.R
\name{run_one_day}
\alias{run_one_day}
\title{Runs SPLASH at a single location for one day}
\usage{
run_one_day(lat, elv, n, y, wn, sf, tc, pn, kCw = 1.05, kWm = 150)
}
\arguments{
\item{lat}{double, decimal degrees.}

\item{elv}{double, elevation, m A.S.L.
Default: \eqn{0}.}

\item{n}{double, day of year.}

\item{y}{double, year.
Default: \eqn{0}.}

\item{wn}{double, daily soil moisture content, mm (wn).}

\item{sf}{double, fraction of sunshine hours.
Default: \eqn{1}.}

\item{tc}{double, mean daily air temperature, degrees C.
Default: \eqn{23.0}.}

\item{pn}{double, daily precipitation, mm/day.}

\item{kCw}{double, supply constant, mm/hr.
Default: \eqn{1.05} (Federer, 1982)}

\item{kWm}{double, soil moisture capacity, mm.
Default: \eqn{150} (Cramer-Prentice, 1988)}
}
\value{
List with the following components:
\itemize{
\item ho .......... daily solar irradiation, J/m2
\item hn .......... daily net radiation, J/m2
\item ppfd ........ daily PPFD, mol/m2
\item cond ........ daily condensation water, mm
\item eet ......... daily equilibrium ET, mm
\item pet ......... daily potential ET, mm
\item aet ......... daily actual ET, mm
\item wn .......... daily soil moisture, mm
\item ro .......... daily runoff, mm
}
}
\description{
Runs SPLASH at a single location for one day
}
