% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evap.R
\name{sat_slope}
\alias{sat_slope}
\title{Calculate the temperature-dependent slope}
\usage{
sat_slope(tc)
}
\arguments{
\item{tc}{double, air temperature, degrees C.}
}
\value{
double, Pa/K.
}
\description{
This function calculates the temperature-dependent slope of the saturation
pressure temperature curve using the methodology presented in the eMast
energy.cpp script.
}
\references{
Allen, R.G., Pereira, L.S., Raes, D. and Smith, M., 1998. Crop
evapotranspiration-Guidelines for computing crop water requirements-FAO
Irrigation and drainage paper 56. Food and Agriculture Organization of the
United Nations, Rome, 300(9), p.D05109. Available:
http://www.fao.org/docrep/x0490e/x0490e07.htm

Prentice, I.C., Sykes, M.T. and Cramer, W., 1993. A simulation model for the
transient effects of climate change on forest landscapes. Ecological
modelling, 65(1-2), pp.51-70. doi:10.1016/0304-3800(93)90126-D
}
\keyword{internal}
