% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash.R
\name{spin_up}
\alias{spin_up}
\title{Calculate daily totals}
\usage{
spin_up(mdat, dtot)
}
\arguments{
\item{mdat}{list with meteorological data (see the details section).}

\item{dtot}{list with daily totals (see the details section).}
}
\value{
list, daily totals
}
\description{
Calculate daily totals updating the soil moisture until equilibrium.
}
\details{
The list with meteorological data, \code{mdat}, should have the
following components:
\itemize{
\item num_lines ..... double, length of meteorol. variable lists
\item lat_deg ....... double latitude (degrees)
\item elv_m ......... double, elevation (m)
\item year .......... double, year
\item sf ............ list, fraction of sunshine hours
\item tair .......... list, mean daily air temperature (deg. C)
\item pn ............ list, precipitation (mm/d)
}

The list with daily totals, \code{dtot}, should have the following component:
\itemize{
\item wm ............ list, daily soil moisture (mm)
}
}
