% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash-package.R
\docType{package}
\name{splash-package}
\alias{splash}
\alias{splash-package}
\title{splash: Simple Process-Led Algorithms for Simulating Habitats}
\description{
This program calculates bioclimatic indices and fluxes (radiation, 
    evapotranspiration, soil moisture) for use in studies of ecosystem function, 
    species distribution, and vegetation dynamics under changing climate 
    scenarios. Predictions are based on a minimum of required inputs: latitude, 
    precipitation, air temperature, and cloudiness. 
    Davis et al. (2017) <doi:10.5194/gmd-10-689-2017>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/villegar/splash/}
  \item \url{https://splash.robertovillegas-diaz.com/}
  \item \url{https://bitbucket.org/labprentice/splash/}
  \item Report bugs at \url{https://github.com/villegar/splash/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@reading.ac.uk} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Authors:
\itemize{
  \item Tyler W. Davis (\href{https://orcid.org/0000-0003-4312-919X}{ORCID})
  \item Iain Colin Prentice \email{c.prentice@imperial.ac.uk} (\href{https://orcid.org/0000-0002-1296-6764}{ORCID})
  \item Benjamin D. Stocker (\href{https://orcid.org/0000-0003-2697-9096}{ORCID})
  \item Rebecca T. Thomas
  \item Rhys J. Whitley
  \item Han Wang (\href{https://orcid.org/0000-0003-2482-1818}{ORCID})
  \item Bradley J. Evans
  \item Angela V. Gallego-Sala
  \item Martin T. Sykes
  \item Wolfgang Cramer (\href{https://orcid.org/0000-0002-9205-5812}{ORCID})
}

}
\keyword{internal}
