% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash.R
\name{splash}
\alias{splash}
\title{Splash}
\usage{
splash(
  mojo_file_path,
  output_file = "index.Rmd",
  overwrite = TRUE,
  title = NULL,
  author = NULL,
  date = NULL,
  rmd_output = "html_document",
  runtime = "shiny"
)
}
\arguments{
\item{runtime}{}
}
\description{
Splash
}
