% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{berger_tls}
\alias{berger_tls}
\title{Calculate true anomaly and true longitude}
\usage{
berger_tls(n, N, ke = 0.0167, keps = 23.44, komega = 283)
}
\arguments{
\item{n}{Numeric, day of year.}

\item{N}{Numeric, days in a year.}

\item{ke}{Numeric, eccentricity of Earth's orbit.
Default: 0.01670, 2000CE (Berger 1978).}

\item{keps}{Numeric, obliquity of Earth's elliptic.
Default: 23.44, 2000CE (Berger 1978).}

\item{komega}{Numeric, longitude of the perihelion, degrees.
Default: 283, 2000CE (Berger 1978).}
}
\value{
True anomaly and true longitude for a given day.
}
\description{
Calculate true anomaly and true longitude
}
\references{
Berger, A. L. (1978), Long term variations of daily insolation
and quaternary climatic changes, J. Atmos. Sci., 35, 2362-2367.
}
